// src/api/models/RoomSessionInput.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type RoomSessionInput = {
    /**
     * ID of the room
     */
    room_id: number;
    /**
     * ID of the session
     */
    session_id: number;
    /**
     * Room session start time
     */
    start_time: string;
    /**
     * Room session end time
     */
    end_time: string;
};