// src/api/models/Room.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Room = {
    /**
     * Room ID
     */
    id: number;
    /**
     * Room name
     */
    name: string;
    /**
     * Room capacity (number of people)
     */
    capacity: number;
    /**
     * Timestamp when the room was created
     */
    created_at: string;
    /**
     * Timestamp when the room was last updated
     */
    updated_at: string;
};